parametri bloka mogu biti float

+++ csmp_blok.py
@@ -6,9 +6,9 @@ class CSMPBlok():
     ulaz1:int = -1
     ulaz2:int = -1
     ulaz3:int = -1
-    par1:int = -1
-    par2:int = -1
-    par3:int = -1
+    par1:float = -1
+    par2:float = -1
+    par3:float = -1
     sifra_bloka:int = -1
     rb_bloka:int = -1
     sortiran:int = False
